requirements.txt
asgiref==3.4.0
Django==3.2.4
django-active-link==0.1.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-taggit==1.5.1
django-widget-tweaks==1.4.8
Pillow==8.3.0
pytz==2021.1
sqlparse==0.4.1
